feat(dashboard): insights v2 page
