+++ frontend/components/dashboard/document-list-card.tsx
@@ -84,9 +84,9 @@ export function DocumentListCard({
             {/* Card Header */}
             <div className={cn(
                 "px-6 py-4 border-b border-gray-100 bg-gray-50 rounded-t-2xl relative z-20",
-                variant === 'flat' && "bg-transparent border-none px-0 py-2 mb-2",
+                variant === 'flat' && "bg-transparent border-none px-0 py-0 mb-0",
                 // If hiding title, we create a 'toolbar' feel
-                hideTitle && "flex justify-end pt-0 pb-2 border-b-0"
+                hideTitle && "flex justify-end pt-0 pb-0 border-b-0"
             )}>
                 <div className={cn(
                     "flex items-center justify-between mb-3 w-full",
@@ -218,7 +218,7 @@ export function DocumentListCard({
             )}>
                 {isLoading ? (
                     Array.from({ length: 5 }).map((_, i) => (
-                        <div key={i} className="px-6 py-5 flex items-center gap-4 animate-pulse border-b border-gray-50 last:border-0">
+                        <div key={i} className="px-6 py-3.5 flex items-center gap-4 animate-pulse border-b border-gray-50 last:border-0">
                             <div className="h-10 w-10 bg-gray-100 rounded-xl flex-shrink-0"></div>
                             <div className="flex-1 space-y-2">
                                 <div className="h-4 bg-gray-100 rounded w-3/4"></div>
@@ -233,7 +233,7 @@ export function DocumentListCard({
                     </div>
                 ) : (
                     (enableFilter ? filteredFiles : files).map((file, index) => (
-                        <div key={`${file.id}-${file.lastAction || index}`} className="group relative flex items-center gap-4 px-6 py-5 hover:bg-gray-50/80 transition-all">
+                        <div key={`${file.id}-${file.lastAction || index}`} className="group relative flex items-center gap-4 px-6 py-3.5 hover:bg-gray-50/80 transition-all">
 
                             {/* Rank Badge */}
                             {showRank && (
